Add files via upload
These are the two most important code files for our project. userinterface.py is the one that works less well. It tries to calculate an equation thousands of times to graph the results. It should be a curve, but instead it is a straight line. We think we know what causes it, but we don't know how to fix it.
The other one, setuserinterface.py, just calculates it once, but this works better. We don't know why one works well and the other doesn't.